refactor: move processField to its own file
